refactor: reorganize Trash module to align with project architecture
- Move Trash view components from src/app/views/Trash to src/views/Trash following view-based structure
- Relocate Trash services from view folder to src/services/Trash following separation of concerns
- Move useTrashPagination hook from view to src/hooks for better reusability
- Update all import paths across the codebase to reflect new structure
- Consolidate Trash component exports in new index file
- Convert all service imports to use absolute app/ paths for consistency
