+++ src/app/drive/components/DriveExplorer/DriveExplorerItem/hooks/useDriveItemActions.tsx
@@ -1,7 +1,7 @@
 import { createRef, useMemo } from 'react';
 
 import navigationService from 'app/core/services/navigation.service';
-import moveItemsToTrash from 'app/views/Trash/services/move-items-to-trash';
+import moveItemsToTrash from '../../../../../../services/Trash/move-items-to-trash';
 import {
   getDatabaseFilePreviewData,
   updateDatabaseFilePreviewData,
