+++ src/app/drive/components/NameCollisionDialog/NameCollisionContainer.tsx
@@ -1,7 +1,7 @@
 import { FC, useEffect, useMemo, useState } from 'react';
 import { connect } from 'react-redux';
 import NameCollisionDialog, { OPERATION_TYPE, OnSubmitPressed } from '.';
-import moveItemsToTrash from '../../../views/Trash/services/move-items-to-trash';
+import moveItemsToTrash from '../../../../services/Trash/move-items-to-trash';
 import errorService from '../../../core/services/error.service';
 import { RootState } from '../../../store';
 import { useAppDispatch, useAppSelector } from '../../../store/hooks';
