--- src/app/views/Trash/index.ts
@@ -1,9 +1,0 @@
-export { default } from './Trash';
-export { default as ClearTrashDialog } from './components/ClearTrashDialog/ClearTrashDialog';
-export { default as StopSharingAndMoveToTrashDialogWrapper } from './components/StopSharingAndMoveToTrashDialog/StopSharingAndMoveToTrashDialogWrapper';
-export { default as StopSharingAndMoveToTrashDialog } from './components/StopSharingAndMoveToTrashDialog/StopSharingAndMoveToTrashDialog';
-export { useTrashPagination } from './hooks/useTrashPagination';
-export { getTrashPaginated, getWorkspaceTrashPaginated } from './services/get_trash';
-export { default as clearTrash } from './services/clear-trash';
-export { default as deleteItems } from './services/delete-items';
-export { default as moveItemsToTrash } from './services/move-items-to-trash';
