+++ src/services/Trash/batch-processor.ts
