+++ src/services/Trash/clear-trash.ts
@@ -1,10 +1,10 @@
 import { t } from 'i18next';
-import { SdkFactory } from '../../../core/factory/sdk';
-import errorService from '../../../core/services/error.service';
-import workspacesService from '../../../core/services/workspace.service';
-import notificationsService, { ToastType } from '../../../notifications/services/notifications.service';
-import { store } from '../../../store';
-import { storageActions } from '../../../store/slices/storage';
+import { SdkFactory } from 'app/core/factory/sdk';
+import errorService from 'app/core/services/error.service';
+import workspacesService from 'app/core/services/workspace.service';
+import notificationsService, { ToastType } from 'app/notifications/services/notifications.service';
+import { store } from 'app/store';
+import { storageActions } from 'app/store/slices/storage';
 
 const clearTrash = async (workspaceId?: string): Promise<void> => {
   let deletingItemsToastId;
