+++ src/services/Trash/get_trash.test.ts
@@ -1,37 +1,37 @@
 import { describe, it, expect, vi, beforeEach, Mock } from 'vitest';
 import { getTrash, getTrashPaginated, getWorkspaceTrashPaginated } from './get_trash';
-import { SdkFactory } from '../../../core/factory/sdk';
-import errorService from '../../../core/services/error.service';
-import localStorageService from '../../../core/services/local-storage.service';
-import workspacesService from '../../../core/services/workspace.service';
-import { store } from '../../../store';
-import { storageActions } from '../../../store/slices/storage';
+import { SdkFactory } from 'app/core/factory/sdk';
+import errorService from 'app/core/services/error.service';
+import localStorageService from 'app/core/services/local-storage.service';
+import workspacesService from 'app/core/services/workspace.service';
+import { store } from 'app/store';
+import { storageActions } from 'app/store/slices/storage';
 
-vi.mock('../../../core/factory/sdk', () => ({
+vi.mock('app/core/factory/sdk', () => ({
   SdkFactory: {
     getNewApiInstance: vi.fn(),
   },
 }));
 
-vi.mock('../../../core/services/error.service', () => ({
+vi.mock('app/core/services/error.service', () => ({
   default: {
     reportError: vi.fn(),
   },
 }));
 
-vi.mock('../../../core/services/local-storage.service', () => ({
+vi.mock('app/core/services/local-storage.service', () => ({
   default: {
     getB2BWorkspace: vi.fn(),
   },
 }));
 
-vi.mock('../../../core/services/workspace.service', () => ({
+vi.mock('app/core/services/workspace.service', () => ({
   default: {
     getTrashItems: vi.fn(),
   },
 }));
 
-vi.mock('../../../notifications/services/notifications.service', () => ({
+vi.mock('app/notifications/services/notifications.service', () => ({
   default: {
     show: vi.fn(),
   },
@@ -44,13 +44,13 @@ vi.mock('../../../notifications/services/notifications.service', () => ({
   },
 }));
 
-vi.mock('../../../store', () => ({
+vi.mock('app/store', () => ({
   store: {
     dispatch: vi.fn(),
   },
 }));
 
-vi.mock('../../../store/slices/storage', () => ({
+vi.mock('app/store/slices/storage', () => ({
   storageActions: {
     clearSelectedItems: vi.fn(),
     setItemsOnTrash: vi.fn(),
