+++ src/services/Trash/get_trash.ts
@@ -1,12 +1,12 @@
 import { t } from 'i18next';
-import { SdkFactory } from '../../../core/factory/sdk';
-import errorService from '../../../core/services/error.service';
-import localStorageService from '../../../core/services/local-storage.service';
-import workspacesService from '../../../core/services/workspace.service';
-import { DriveItemData } from '../../../drive/types';
-import notificationsService, { ToastType } from '../../../notifications/services/notifications.service';
-import { store } from '../../../store';
-import { storageActions } from '../../../store/slices/storage';
+import { SdkFactory } from 'app/core/factory/sdk';
+import errorService from 'app/core/services/error.service';
+import localStorageService from 'app/core/services/local-storage.service';
+import workspacesService from 'app/core/services/workspace.service';
+import { DriveItemData } from 'app/drive/types';
+import notificationsService, { ToastType } from 'app/notifications/services/notifications.service';
+import { store } from 'app/store';
+import { storageActions } from 'app/store/slices/storage';
 
 const getTrash = async (): Promise<void> => {
   const trashClient = SdkFactory.getNewApiInstance().createTrashClient();
