+++ src/views/Trash/StopSharingAndMoveToTrashDialogWrapper.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { MAX_SHARED_NAME_LENGTH } from '../../../../share/views/SharedLinksView/SharedView';
+import { MAX_SHARED_NAME_LENGTH } from 'app/share/views/SharedLinksView/SharedView';
 import StopSharingAndMoveToTrashDialog from './StopSharingAndMoveToTrashDialog';
-import { sharedThunks } from '../../../../store/slices/sharedLinks';
-import { DriveItemData } from '../../../../drive/types';
+import { sharedThunks } from 'app/store/slices/sharedLinks';
+import { DriveItemData } from 'app/drive/types';
 
 interface StopSharingAndMoveToTrashDialogWrapperProps {
   selectedItems: any[];
