+++ src/views/Trash/index.ts
@@ -1,0 +1,9 @@
+export { default } from './Trash';
+export { default as ClearTrashDialog } from './ClearTrashDialog';
+export { default as StopSharingAndMoveToTrashDialogWrapper } from './StopSharingAndMoveToTrashDialogWrapper';
+export { default as StopSharingAndMoveToTrashDialog } from './StopSharingAndMoveToTrashDialog';
+export { useTrashPagination } from '../../hooks/useTrashPagination';
+export { getTrashPaginated, getWorkspaceTrashPaginated } from '../../services/Trash/get_trash';
+export { default as clearTrash } from '../../services/Trash/clear-trash';
+export { default as deleteItems } from '../../services/Trash/delete-items';
+export { default as moveItemsToTrash } from '../../services/Trash/move-items-to-trash';
